feat: Simple shell 0.1
The Shell works for simple commands without any arguments.
The Shell was tested using the commands "/bin/ls" and "/bin/pwd".
It is possible to exit the shell using the shortcut Ctrl + D.

bug - The exit command doesn't work properly.
It runs only when used first or twice in a row. If it's preceeded
by another input the program just ignores it. The bug might be caused
by the child process.